remove ookla_speedtest.py + rename file ookla_speedtest_cli.py -> ookla_speedtest.py
